refactor(lib/vec/types): rename index.d.ts to types.ts
